+++ docs/release-notes.md
@@ -14,6 +14,9 @@ nav_order: 9
 
 ### Misc. changes
 
+- Warn on root partition size is too small _(fcos 1.3.0-1.8.0-exp)_
+- Warn on root partition constrained by another partition _(fcos 1.3.0-1.8.0-exp)_
+
 ### Docs changes